Include all snapshots and move them to a separate snapshot folder
